Modified feature toggles.
There is a new feature toggle for a modified parameter passing.
In the new form, some parameters are passed as a file. Currently, only
Linux targets are supported, as "export" is used to make the variables
available globally.

The current binary is updated and the wrapper script imports the
parameters file if it is set :)