Modify ConfigurationFactory and pull over the dynamic class creation code to the LibrariesFactory.
Modify Tool Entry to contain a generic file type parameter for file
groups.

Add the brawl workflow code to provide the _assembleLoadFilesCall which
looks like:
set inputfiles = loadfilesWith "${LOAD_FASTQ_FILES}"()
This calls a script on the remote computer and gets a file group out of
it. The script is configured like a tool. Also a test for this method is
available.

Extend the dynamic class creation code so it does create file groups
now.