Changed obscure Null pointer exception with null message.
Summary:
We use Reflection to get the proper constructor for the command factory class. It seems, that the wrong constructor was retrieved, because the order of the constructors got messed up. I believe, that it
has to do with Groovy 2.4.x but I am not totally sure about it. I fixed the code, so that a specific constructor will be used.

Test Plan: Unfortunately none. Repeated startup should show, if the error was really solved.

Reviewers: kaercher, kensche

Differential Revision: https://tbi-cb14.inet.dkfz-heidelberg.de/D125