Fixed rerun and compilation issues.
Summary:
ExecutionContext object was not properly set when calling setExecutionContext in class FileObject
Fix a bug in FileObject: setExecutionContext did not set the newContext object but used a cyclic reference. This prevented rerun in some cases.

Fix a bug in roddy compilation (and recompile) where the JDK version
could not be retrieved when the jar file was missing (and to produce the
jar file is the task of the compile step. :) )

Test Plan: Manual

Reviewers: biegm

Reviewed By: biegm

Differential Revision: https://tbi-cb14.inet.dkfz-heidelberg.de/D168