Initial version for the new autoselect feature.
Summary:
Roddy is now able to chose the right application version based on three
different modes:
2 fast modes (bash) with the --usePluginVersion option set on command
line or in ini.
1 slower mode using Roddy current (or the latest available). Guess we
have to talk more about current / latest as it might not always be
available.

Test Plan: First review, No unit tests... Florian, please also take a look at it! This is an important and new feature.

Reviewers: kaercher, kensche

Reviewed By: kensche

Differential Revision: https://eilslabs-phabricator.dkfz.de/D253