Introduce a new feature for job output parameters
This commit changes the behaviour of the boolean "check" for job output file
checks. Instead of a simple boolean, we now will have a more
sophisticated conditional class. This can either be a pure boolean OR a
conditional.

Syntax is check="conditional:VARNAME"
or
check="true|false"

Errors in the definition will all fall back to true!