Quickfix recursive loop error; Introduce class for Roddy integration tests
The recursive loop error, which appeared after moving code to
RuntimeService is now done with a quick fix. However, code should be
refactored and the detail loading code, which is responsible for the
problem, needs to be moved to another point.

Introduce a first (empty) class for future integration tests of the
compiled Roddy binary. We need this to make sure, that all (important)
options of Roddy will work.