Bugfix for FG output, Cleanup and enhance FilenamePattern(Test)
The wrong parameter name was used for file group outputs. The correct
one is now set.

Some code was moved in FilenamePattern and Job to make the code more
readable.

Extracted methods in FilenamePattern to create better readable code
(Especially the apply method)

Created some tests to ensure that nothing breaks and to document
features.

Added the possibility for short variable syntax in filename patterns:
Both refer to the same variable: "${cvalue,name="test"}_${test}"
The short syntax does not allow default values. It is either set or
left like it is.