Clean up startup messages, correct cli toggle input
It is not only about colours. But the testrun/rerun where cleaned up.

Also toggle values are now parsed in a correct way. If a toggle value
does not exist, an error is shown and a list of available toggles is
printed.

If the ini file is not found, Roddy will not start!