Chgd log behaviour, +/- messages, + startup checks, add --c to bash
Added the --c option to the startup scripts.

Changed the logger behaviour so, that it will now be correctly
initialized.
The log directory was wrong. As soon as the class is setup, the correct
directory is set.

Changed colour setting for auto file names. They will be displayed in
another colour.

Removed obsolote or non-trivial / misinterpretable log messages which
did not help.

Added the .checkFilesPreparedForExecution method to ExecutionService
The method will check all files created in writeFilesForExecution on
accessibility.
If strict mode is enabled, more checks will lead to app exit.

Add logger messages to BaseFile.getAvailableFilenames()
The amount and a list of found patterns for a file class will be
printed.

Updated RoddyToolLib to new build.