Add further executability checks for tools
Modify Analysis.executeRun so, that:
- the copied files are checked
- the copied analysis tools are check
Both are checked and errors will be presented.

Add fileIsExecutable to ExecutionContext

Add checkCopiedAnalysisTools to ExecutionService
The method loops over all configured tools and checks, if they are
executable.

Next step would be to mark them as executable, but this is not yet
available.
An error message was implemented. Added a method stub.

Made createJobName in JobManager facade static. There is a
IncompatibleClassChangeError
thrown, when the createJobName method is called.