Improve some code, Fix bugs (e.g. onScriptParm pattern)
Add EmptyResourceSet class to reflect unset resource set.

Add FToggle: FailOnAutoFilenames, which is applied in strict mode only.
If the toggle is not active, a context error will be created.

Add test code and configuration for toggles based on script parameters

Corrected and simplified findFilenameFromOnScriptParameterPatterns which
returned the same name for all files of the same type within a tuple.