Rework the NativeWorkflowInterface
Add the getFolderForConvertedNativePlugins method to Roddy class.

Uncomment the NativeWorkflow class.

Introduce the PluginDirectoryInfo enum which contains very basic
information about a plugin directory:

Introdcue the PluginType enum which is one
- NONE 		- Marks invalid plugins
- NATIVE 	- Marks a plugin as Native
- RODDY 	- A regular Roddy workflow
- SNAKEMAKE - A snakemake based workflow. Let's see if we keep it.

Introduce the convertNativePluginsIfNeccessary method in
LibrariesFactory.

Add some getters and setters to PluginInfo class. This is neccessary
because of static type checking in another method.