Two (urgent) fixes
Fix one: Extend the help output

Fix two: Without BE there was a position in ExecutionContext where a job
result was checked for being null. With BE we used the result as an
input for the new JobResult wrapper object. However this was not
properly check and lead to a NullPointerException