Add multi-user support for GroovyServ
Rename downloadFunctions to networkFunctions

Add getExistingOrNewGroovyServPort() function, which seeks for an active
user instance of GroovyServ and tries to extract the used port or, if
that fails, tries to find a free port.

Add FindPort.java and .class file. This has a single method which gets a
free port from the underlying system.