Add full per job configuration files
With job parameter files, the Roddy internal mechanism was not working
properly when it came to the job / bash level.

Removed most of the blacklist values which where necessary for parameter
conversion in earlier Roddy versions.

Moved job configuration file store method to the Job class.

Add a run flag to the test workflow, which can let you run it with or
without errors.

Add some more entries to the test workflow, which lets you test
dependency issues.