Add some methods to make file loading easier
At some point we implemented the file creation helper object classes
which are good to use in a lot of occassions. However, the mechanism is
somewhat "complex" and takes a lot of code.

As we already have some shortcut methods for method calling in the
workflow, we also will feature these shortcut methods for file loading.