Change job config storage behaviour
Moved the line where the store job config is called to a location where
it is actually used. The job configuration is only necessary, when jobs
are run. In case of QUERY_STATUS, the line won't work an an error is
raised, that the file cannot be written. That is, because the exec
folder is not available at that point in time.

Also change one of the test workflows a bit so it will run again.