Fix wrong value bundle id during conversion bug
After some of the last changes, we corrected the behaviour of the
recursive value containers. The new code relied more on e.g. the method
getId() of contained values (of type Identifiable). However, when value
bundles were converted to Bash, the id was an integer starting from 0.

This was, because the value bundles did not store their string id but an
incrementing value! The behaviour has been fixed and the bundles are now
converted properly.