Remove confusing exec dir message
As of one update, Roddy will create more than one execution directory
(e.g. to have all scripts on the target system, when using a tool to
load files). This directory will always be removed and the message
telling the user, that the directory was created, is unnecessary.
Therefore, disable the message, if the context is not allowed to submit
jobs.