Implement testrun and run for Brawl workflows
TODO: Inline scripts are still not generated
TODO: Variables and configuration values are not yet possible

Add loadAndPreprocessBrawlFile to ConfigurationFactory. The method
creates a basic xml for the loaded Brawl workflow.

Change explicit in BrawlWorkflow to be a method. The method now assigns
the passed closure to the explicitWorkflow variable. The closure is
called when execute is called.

Rename the call methods in BrawlWorkflow to run. Whyever it is but
Groovy does not like the call syntax as this somehow (not reproducable
in other tests) messes up with the closure call method.

Load Brawl workflow files as configuration files.

Extend the Brawl workflow test to run a test workflow without Roddy.

Adapted the test workflow in the test plugin.