Fix inline scripts for Brawl DSL
Inline scripts are now created in the target execution directory.

Make ExecutionService.copyAnalysisToolsForContext public, so it can be
accessed by e.g. BrawlWorkflow.

Modify BrawlWorkflow so, that the target analysis tools directory is
removed and recreated on run. This is necessary, as Brawl scripts are
converted at a later point of runtime than normally and would otherwise
not be available for execution. We could keep the tools directory but
this would need a lot of code adaption. E.g. Just copying the tools
again would lead to directories like
analysisTools/inlineScripts/inlineScripts. Recreation is easier to
understand (two lines of code).