Cleaned up catches (removed, more specific) and improved error messages if no filename patterns match.
Now always an exception is thrown and caught to be added into a ExecutionContext.addErrorEntry.

Note that at BaseFile creation time, because here already the getFilename method is called which in turn "applies" patterns to the BaseFile, the currently executed tool and matching (!) filename patterns are needed. Therefore lots of tests needed to be adapted.