Introduce the ConfigurationIssue system
Add the IConfigurationIssueInterface, and the ConfigurationIssue class
and the ConfigurationIssueLevel and ConfigurationIssueTemplate
enumerations.
- Configuration issues are stored for configuration objects and for
  configuration value objects.
- They are either of type info, warning or error.
- Issues in configuration values are passed to the configuration on
  request, they are not stored there! This has changed to previous
versions of Roddy.

Implement the IConfigurationIssueInterface interface in Configuration,
ConfigurationValue.
- The interface adds the methods: hasWarnings/Errors, addWarning/Error,
  getWarnings/Errors.
- getWarning/Error returns a copy of the underlying collections in its
  current implementations.

Move ConfigurationTest to ConfigurationSpec, add a test for error and
warning passthrough to the configuration object.

Configuration class:
- Rename Configuration.hasErrors to hasLoadErrors.
- Groovyfy some code in the Configuration class.

Analysis class:
- Rename Analysis.java to Analysis.groovy and Groovyfy it.
- Rebuild and extend the print errors and warning methods in Analysis
  class to match the new possibilites and make the code more clear.

Changes to ConfigurationValue
- Add constructor methods for int, float, double, bool to
  ConfigurationValue.
- Enable auto-validation of ConfigurationValue objects upon their
  construction.
- Groovyfy some the class.

The Enumeration class has the method getDefaultCValueTypeEnumeration()
now. The method will return a default list of available configuration
value types and their default validators.
Also did a code cleanup to Groovyfy it.

Move ConfigurationValueValidator.java to
ConfigurationValueValidator.groovy and cleanup the class.
Implement the IConfigurationIssueInterface and add proper fields and
methods.

Rename DefaultValidator.java to DefaultValidator.groovy and Groovyvy
Extend the DefaultValidator class with checkDollarSignUsage(). The
method now stores its errors and warnings in their fields in the parent
class.

Add test for proper dollar character usage to the ConfigurationFactory
test.

Add ConfigurationIssueSpec test class.

Add the DefaultValidatorSpec test class.

Remove the ConfigurationTest class.