Fixed ConfigurationIssue issues
Added a new constructor for elevated values to the ConfigurationValue
class. This also speeds up the code, as the validation result will just
be taken from the parent value. Otherwise, an unnecessary revalidation
would occur and they are quite expensive.

Changed the toString of ConfigurationIssue to print out the message
instead of some made up text.

Added equals and hashCode to EnumerationValue to make them comparable.

Add printErrorsAndWarnings() to Analysis to allow a single call to print
out warnings, info messages and errors.

Add detailed warnings and error output for condensed messages to the
detailed logfile.

Added more tests.