Fixed issues with RoddyTestSpec
Added a comment to Analysis.getConfiguration(). The method will fail for
many tests, when the tests do not also initialize the ExecutionService
and FileSystemAccessProvider. The comment describes the method which
needs to be added to the test class.

Added the aforementioned method to ContextResourceTest

Changed ContextResource so, that the contextResource object is now
@Shared and a @ClassRule but NOT static! Also the static initializer in
the anonymous creation is removed.