Add an enumeration to store exit codes and descriptions/messages
Added the ExitReasons enumeration which keeps entries for knwon and
unknown Roddy exit codes. Each value in the enum stores a short message
and the exit code itself. The message may be used, however its usage
is not enforced. E.g. bigger error messages with help texts or custom
messages with variable parts are better printed in another way.

Added a table about the exit codes and a bit of more description to the
FAQ file in the read the docs files.

Changed all calls of Roddy.exit(int) to use the exit codes of the
entries in ExitReasons.