Add first version of JSON configuration file loader and tests
Add the getResourceFile() method to the RoddyTestSpec class. This makes
import of resource files easy.

Add the test method testReadProcessingTools() to the
ConfigurationFactoryTest class. This is unfinished in this commit.

In ConfigurationFactory class:
- Add the loadJSONFile() and loadYAMLFile() method to the
- Add the selectProcessingtoolsNode() method to allow the lower case
  <processingtools> entry instead of <processingTools> in xml
  configuration files.

Add the JSONConverter class which can load a file and convert the string
to a valid Roddy xml configuration file. This way, we can facilitate our
existing xml checks and still use JSON and later YAML functionality.

Add the JSONConverterTest class and example configuration files. This
covers project and analsis xml configuration files. The test method will
run on a line by line test because a full text comparison did not work.