Add YAML cfg files, fixes
Add the availableanalyses and processingtools tag to the
projectConfigurationValidation.xst file.

Change BashConverter to write availableanalyses instead of
availableAnalyses.

Move the configuration conversion code from JSONConvert to the
MapConverter class. All methods work on maps and can be reused for YAML
configuration files.

Added the YAMLConverter class and make it extend MapConverter.

Add the MapConverter class and move in code from JSONConverter. Extend
and fix the code where necessary. E.g. subconfigurations are now parsed.

Add JSON and YAML loading to the ConfigurationFactory.

Add the selectAvailableAnalysesNode method and fix the
selectProcessingToolsNode method. Both use the size() method to check
if either the new form or the old form is used. Fix the test for the
processingTools method.

Fix, rename and extend the test files for the JSON/YAML/MapConverter
classes.

Fix the test classes for JSON and YAMLConverter. Both work a lot more
reliable now, though they need to be quite complicated.

Include Snakeyaml dependency in build.gradle.