Fix a bug which made CValue.toFile() output wrong paths
When dealing with multi-inherited configuration values, the method
failed and put the Roddy application directory in front of the final
value.

To fix this, the toFile() method will now work with File.path() instead
of File.absolutePath(). I tested it with four different workflows in a
productive environment using printidlessruntimeconfig.