Fixed buggy warning for cvalue validation with subsequent variable references.
The RE consumed the values before and could only match the first of "${a}${b}".

Simplified the expression by factoring out the suffix (the variable reference) and implemented positive lookbehind to deal with overlap of patterns (the lookbehind now can match in the "${a}" and thus count both "${a}" and "${b}" as variable references.